refactor(saves): saves package __init__.py is re-export only
Move the SaveService façade class definition from saves/__init__.py
to saves/facade.py. The package __init__.py now holds only the
package's contract docstring and a re-export of SaveService.

Same convention as sub-package __init__.py (status/, sync_engine/) —
init = namespace marker + public-API re-export, primary code in
named modules. Applies the rule uniformly: the saves package's
public surface is one line in __init__.py.

External callers continue to import unchanged
(`from services.saves import SaveService`); the import resolves
through the new __init__.py re-export. Internal sub-services that
TYPE_CHECKING-back-ref `services.saves.SaveService` still work for
the same reason.

Refs #257.
