+++ py_modules/services/saves/__init__.py
@@ -1,1800 +1,12 @@
-"""SaveService façade — composes the save-sync sub-services and exposes
-the public callable API for save sync.
+"""Save-sync subsystem.
 
-Business logic and state live in dedicated sub-modules; this file holds
-composition + thin delegates. RomM communication goes through
-``RommApiProtocol``; no ``import decky`` (error helpers come from
-``lib.errors``).
+The package's public API is the ``SaveService`` façade — composes the
+save-sync sub-services (state, sync_engine, status, versions, slots)
+and exposes the callable surface consumed by the Decky entrypoints.
+RomM communication goes through ``RommApiProtocol``; no ``import decky``
+(error helpers come from ``lib.errors``).
 """
 
-from __future__ import annotations
+from services.saves.facade import SaveService
 
-import asyncio
-import contextlib
-import hashlib
-import os
-import socket
-from dataclasses import asdict
-from datetime import UTC, datetime
-from typing import TYPE_CHECKING, ClassVar
-
-from models.saves import SaveConflict
-
-from domain.emulator_tag import build_emulator_tag, detect_core_change
-from domain.save_extensions import get_save_extensions
-from domain.save_path import resolve_save_dir
-from lib.errors import classify_error
-from lib.iso_time import parse_iso_to_epoch
-from services.protocols import (
-    CoreNameProviderFn,
-    CoreResolverFn,
-    RetryStrategy,
-    RommApiProtocol,
-    RomsPathProvider,
-    SavesPathProvider,
-)
-from services.saves._helpers import _local_save_target
-from services.saves.state import StateService
-from services.saves.status import StatusService
-from services.saves.sync_engine import SyncEngine
-from services.saves.versions import VersionsService
-
-_DEVICE_NOT_REGISTERED = "Device not registered"
-_NO_MIGRATION = object()  # sentinel: no slot migration requested
-
-if TYPE_CHECKING:
-    import logging
-    from collections.abc import Callable
-
-    from services.protocols import EventEmitter
-
-
-_SYNC_DISABLED_MSG = "Save sync is disabled"
-
-
-class SaveService:
-    """Bidirectional save file sync between local RetroDECK and RomM server.
-
-    Parameters
-    ----------
-    romm_api:
-        Protocol adapter for all RomM save/notes HTTP operations.
-    retry:
-        Retry strategy — provides ``with_retry`` and ``is_retryable``.
-    state:
-        Live reference to the main plugin state dict (``installed_roms``,
-        ``shortcut_registry``).
-    save_sync_state:
-        Live reference to the save-sync state dict.  Caller should
-        pre-populate via :meth:`init_state` / :meth:`load_state`.
-    loop:
-        The plugin's ``asyncio`` event loop (for ``run_in_executor``).
-    logger:
-        Standard-library logger (replaces ``decky.logger``).
-    runtime_dir:
-        Absolute path to the plugin runtime directory (for
-        ``save_sync_state.json`` persistence).
-    get_saves_path:
-        Callable returning the current RetroDECK saves directory.
-    get_roms_path:
-        Callable returning the current RetroDECK roms directory.
-    get_active_core:
-        Callable resolving the active RetroArch core for a system/game.
-        Returns ``(core_so, label)`` tuple; either may be None if unresolved.
-        This is an ES-DE question (``which core runs this ROM?``).
-    get_core_name:
-        Callable returning the RetroArch canonical ``corename`` field from
-        a core's ``.info`` file for a given ``core_so`` (e.g. ``"mgba_libretro"``
-        → ``"mGBA"``). Optional. When ``sort_savefiles_enable`` is active on
-        RetroArch, this is the authoritative name used for the per-core save
-        subdirectory — it is NOT the same as the ES-DE UI label returned by
-        ``get_active_core`` (see the Config-Source-Parsers wiki page for the
-        one-parser-per-source rationale). When ``None`` or when resolution
-        fails at runtime, SaveService warns and falls back to the parent
-        directory path; see ``_resolve_retroarch_corename``.
-    detect_sort_change:
-        Optional synchronous callback that refreshes save-sort state from
-        the live RetroArch config (wired to
-        ``MigrationService.detect_save_sort_change`` in ``bootstrap``).
-        Save-sync MUST see fresh save-sort state before computing
-        ``saves_dir`` — otherwise a direct-Steam-launch with no pre-launch
-        detect trigger would silently download stale server content to the
-        wrong layout and destroy real user progress during the subsequent
-        migration (#238). ``pre_launch_sync`` and ``post_exit_sync`` invoke
-        this callback once at their entry point. ``None`` disables the
-        call (used only in unit tests where state is seeded explicitly);
-        failures are logged and swallowed so save-sync degrades
-        gracefully to the previously-known state.
-    """
-
-    _LOG_LEVELS: ClassVar[dict[str, int]] = {"debug": 0, "info": 1, "warn": 2, "error": 3}
-
-    def __init__(
-        self,
-        *,
-        romm_api: RommApiProtocol,
-        retry: RetryStrategy,
-        settings: dict,
-        state: dict,
-        save_sync_state: dict,
-        loop: asyncio.AbstractEventLoop,
-        logger: logging.Logger,
-        runtime_dir: str,
-        get_saves_path: SavesPathProvider,
-        get_roms_path: RomsPathProvider,
-        get_active_core: CoreResolverFn,
-        get_core_name: CoreNameProviderFn | None = None,
-        plugin_version: str = "0.0.0",
-        emit: EventEmitter | None = None,
-        detect_sort_change: Callable[[], None] | None = None,
-        is_retrodeck_migration_pending: Callable[[], bool] | None = None,
-    ) -> None:
-        self._romm_api = romm_api
-        self._retry = retry
-        self._settings = settings
-        self._state = state
-        self._state_svc = StateService(
-            save_sync_state=save_sync_state,
-            state=state,
-            runtime_dir=runtime_dir,
-            logger=logger,
-        )
-        # Alias the dict so the dozens of self._save_sync_state[...] call
-        # sites elsewhere in SaveService keep working unchanged. Both names
-        # reference the same underlying dict object.
-        self._save_sync_state = self._state_svc.data
-        self._loop = loop
-        self._logger = logger
-        self._get_saves_path = get_saves_path
-        self._get_roms_path = get_roms_path
-        self._get_active_core = get_active_core
-        self._get_core_name = get_core_name
-        self._plugin_version = plugin_version
-        self._emit = emit
-        self._detect_sort_change = detect_sort_change
-        self._is_retrodeck_migration_pending = is_retrodeck_migration_pending
-        # Per-rom lock dict — serializes concurrent sync operations on the
-        # same rom_id (pre_launch_sync, post_exit_sync, manual sync, resolve).
-        self._rom_sync_locks: dict[int, asyncio.Lock] = {}
-        self._versions = VersionsService(
-            save_service=self,
-            state_svc=self._state_svc,
-            romm_api=self._romm_api,
-            logger=self._logger,
-        )
-        self._status = StatusService(
-            save_service=self,
-            state_svc=self._state_svc,
-            romm_api=self._romm_api,
-            logger=self._logger,
-        )
-        self._sync_engine = SyncEngine(
-            save_service=self,
-            state_svc=self._state_svc,
-            romm_api=self._romm_api,
-            retry=self._retry,
-            logger=self._logger,
-        )
-
-    def _rom_lock(self, rom_id: int) -> asyncio.Lock:
-        """Return the lock for this rom_id, creating it lazily."""
-        if rom_id not in self._rom_sync_locks:
-            self._rom_sync_locks[rom_id] = asyncio.Lock()
-        return self._rom_sync_locks[rom_id]
-
-    # ------------------------------------------------------------------
-    # Debug logging helper
-    # ------------------------------------------------------------------
-
-    def _log_debug(self, msg: str) -> None:
-        configured = self._settings.get("log_level", "warn")
-        if self._LOG_LEVELS.get("debug", 0) >= self._LOG_LEVELS.get(configured, 2):
-            self._logger.info(msg)
-
-    def _get_server_device_id(self) -> str | None:
-        """Return the server device ID if registered, else None."""
-        return self._save_sync_state.get("server_device_id")
-
-    # ------------------------------------------------------------------
-    # State Management
-    # ------------------------------------------------------------------
-
-    @staticmethod
-    def make_default_state() -> dict:
-        """Return a fresh default save-sync state dict."""
-        return StateService.make_default_state()
-
-    def init_state(self) -> None:
-        """Populate ``_save_sync_state`` with defaults (idempotent)."""
-        self._state_svc.init_state()
-
-    def _migrate_loaded_state(self) -> None:
-        """Apply schema migrations to data just read from disk."""
-        self._state_svc._migrate_loaded_state()
-
-    def load_state(self) -> None:
-        """Load save sync state from disk, merging with defaults."""
-        self._state_svc.load_state()
-
-    def save_state(self) -> None:
-        """Persist save sync state to disk (atomic write)."""
-        self._state_svc.save_state()
-
-    def prune_orphaned_state(self) -> None:
-        """Remove save sync state entries for rom_ids no longer in shortcut registry."""
-        self._state_svc.prune_orphaned_state()
-
-    # ------------------------------------------------------------------
-    # ROM / path helpers
-    # ------------------------------------------------------------------
-
-    def _resolve_retroarch_corename(self, system: str, rom_filename: str) -> tuple[str | None, str | None]:
-        """Resolve the RetroArch ``corename`` for a system/ROM.
-
-        Asks ES-DE (via ``get_active_core``) **which** core is active for
-        this ROM, then asks the RetroArch ``.info`` parser (via
-        ``get_core_name``) **what** RetroArch calls that core in its own
-        subsystem — which is the authoritative name used for per-core save
-        subdirectories when ``sort_savefiles_enable`` is active.
-
-        One parser per source: the ES-DE label (second element of the
-        ``get_active_core`` tuple) is NOT a valid substitute for the
-        RetroArch corename. See the Config-Source-Parsers wiki page and
-        the reference implementation in ``MigrationService``.
-
-        Returns ``(corename, core_so)``. Either element may be ``None``
-        when resolution fails at that step: ``core_so`` is ``None`` when
-        ES-DE cannot determine the active core, ``corename`` is ``None``
-        when ``.info`` parsing returns nothing (or when ``get_core_name``
-        is not injected). Returning the tuple — rather than just
-        ``corename`` — lets callers include ``core_so`` in diagnostic
-        logs so users can identify which ``.info`` file is at fault.
-        Callers choose their own fallback strategy (e.g. warn and fall
-        back for critical-path SaveService flows; skip and warn for
-        one-shot migrations).
-        """
-        if self._get_core_name is None:
-            return (None, None)
-        core_so, _label = self._get_active_core(system, rom_filename)
-        if not core_so:
-            return (None, None)
-        corename = self._get_core_name(core_so)
-        return (corename or None, core_so)
-
-    def _get_rom_save_info(self, rom_id: int) -> dict | None:
-        """Get save-related info for an installed ROM.
-
-        Returns dict with keys: system, rom_name, saves_dir, platform_slug, file_path
-        or None if not installed.
-        """
-        rom_id_str = str(int(rom_id))
-        installed = self._state["installed_roms"].get(rom_id_str)
-        if not installed:
-            return None
-        system = installed.get("system", "")
-        file_path = installed.get("file_path", "")
-        platform_slug = installed.get("platform_slug", "")
-        if not system or not file_path:
-            return None
-        rom_name = os.path.splitext(os.path.basename(file_path))[0]
-
-        # Use domain save path resolution.
-        # Read sort settings from state (populated by MigrationService at startup).
-        # When a save-sort migration is pending, prefer the *previous* layout:
-        # RetroArch caches its runtime save-path at game-load time, so the
-        # session that just ended still wrote to the old directory. Reading
-        # the current settings here would point sync at the wrong location
-        # and risk downloading stale server content to the new layout (#238).
-        saves_base = self._get_saves_path()
-        roms_base = self._get_roms_path()
-        sort_state = self._pending_sort_settings() or self._state.get("save_sort_settings")
-        if sort_state:
-            sort_by_content = sort_state.get("sort_by_content", True)
-            sort_by_core = sort_state.get("sort_by_core", False)
-        else:
-            sort_by_content, sort_by_core = True, False  # RetroDECK defaults
-
-        # When sort-by-core is active, RetroArch writes per-core subdirs named
-        # by the .info ``corename`` field. Resolve it via the dedicated parser.
-        # See docs: Config-Source-Parsers wiki page ("one parser per source").
-        # Decision: warn-and-fallback (not fail-loud like MigrationService).
-        # SaveService is the critical-path sync flow — every game launch
-        # depends on it. Fail-loud would take down save sync entirely on any
-        # .info hiccup. MigrationService can afford strictness (one-shot),
-        # SaveService cannot (continuous). See issue #232 for history.
-        core_name: str | None = None
-        if sort_by_core:
-            rom_filename = os.path.basename(file_path)
-            core_name, core_so = self._resolve_retroarch_corename(system, rom_filename)
-            if core_name is None:
-                self._logger.warning(
-                    "SaveService: unable to resolve RetroArch corename for "
-                    "%s/%s (core_so=%s) while sort_by_core is enabled. "
-                    "Falling back to the parent save directory, which will "
-                    "not match what RetroArch reads at runtime. Check that "
-                    "the core's .info file is readable under the RetroDECK "
-                    "Flatpak cores directory.",
-                    system,
-                    rom_filename,
-                    core_so if core_so else "unresolved",
-                )
-
-        saves_dir = resolve_save_dir(
-            file_path,
-            saves_base,
-            system,
-            roms_base=roms_base,
-            sort_by_content=sort_by_content,
-            sort_by_core=sort_by_core,
-            core_name=core_name,
-        )
-
-        return {
-            "system": system,
-            "rom_name": rom_name,
-            "saves_dir": saves_dir,
-            "platform_slug": platform_slug,
-            "file_path": file_path,
-        }
-
-    def _pending_sort_settings(self) -> dict | None:
-        """Return previous save-sort settings if a migration is pending, else None.
-
-        Rejects empty dicts to avoid the half-state where ``_get_rom_save_info``'s
-        ``or`` fallback would treat ``{}`` as "no pending migration" (and read
-        current settings) while ``_is_save_sort_changed`` would treat the same
-        ``{}`` as "pending" (and gate sync). Both call sites must agree on
-        what counts as pending — see #238 review finding 3.
-        """
-        prev = self._state.get("save_sort_settings_previous")
-        return prev if prev else None
-
-    def _is_save_sort_changed(self) -> bool:
-        """Check if a save sort migration is pending (detected by MigrationService)."""
-        return self._pending_sort_settings() is not None
-
-    # ------------------------------------------------------------------
-    # File Helpers
-    # ------------------------------------------------------------------
-
-    @staticmethod
-    def _file_md5(path: str) -> str:
-        """Compute MD5 hash of a file."""
-        h = hashlib.md5()
-        with open(path, "rb") as f:
-            for chunk in iter(lambda: f.read(8192), b""):
-                h.update(chunk)
-        return h.hexdigest()
-
-    def _find_save_files(self, rom_id: int) -> list[dict]:
-        """Find local save files for a ROM.
-
-        Returns list of ``{"path": str, "filename": str}``.
-        """
-        info = self._get_rom_save_info(rom_id)
-        if not info:
-            return []
-        rom_name = info["rom_name"]
-        saves_dir = info["saves_dir"]
-        platform_slug = info["platform_slug"]
-        if not os.path.isdir(saves_dir):
-            return []
-        results = []
-        for ext in get_save_extensions(platform_slug):
-            save_path = os.path.join(saves_dir, rom_name + ext)
-            if os.path.isfile(save_path):
-                results.append({"path": save_path, "filename": rom_name + ext})
-        return results
-
-    # ------------------------------------------------------------------
-    # SyncEngine delegates
-    # ------------------------------------------------------------------
-
-    def _get_server_save_hash(self, server_save: dict) -> str | None:
-        """Delegate to :class:`SyncEngine` — see its docstring."""
-        return self._sync_engine._get_server_save_hash(server_save)
-
-    def _update_file_sync_state(
-        self,
-        rom_id_str: str,
-        filename: str,
-        server_response: dict,
-        local_path: str,
-        system: str,
-        *,
-        emulator_tag: str | None = None,
-        core_so: str | None = None,
-    ) -> None:
-        """Delegate to :class:`SyncEngine` — see its docstring."""
-        self._sync_engine._update_file_sync_state(
-            rom_id_str,
-            filename,
-            server_response,
-            local_path,
-            system,
-            emulator_tag=emulator_tag,
-            core_so=core_so,
-        )
-
-    def _do_download_save(self, server_save: dict, saves_dir: str, filename: str, rom_id_str: str, system: str) -> None:
-        """Delegate to :class:`SyncEngine` — see its docstring."""
-        self._sync_engine._do_download_save(server_save, saves_dir, filename, rom_id_str, system)
-
-    def _do_upload_save(
-        self,
-        rom_id: int,
-        file_path: str,
-        filename: str,
-        rom_id_str: str,
-        system: str,
-        server_save: dict | None = None,
-    ) -> dict:
-        """Delegate to :class:`SyncEngine` — see its docstring."""
-        return self._sync_engine._do_upload_save(rom_id, file_path, filename, rom_id_str, system, server_save)
-
-    @staticmethod
-    def _filter_server_saves_to_slot(server_saves: list[dict], active_slot: str | None) -> list[dict]:
-        """Delegate to :class:`SyncEngine` — see its docstring."""
-        return SyncEngine._filter_server_saves_to_slot(server_saves, active_slot)
-
-    @staticmethod
-    def _build_local_input(local_path: str, filename: str) -> dict:
-        """Delegate to :class:`SyncEngine` — see its docstring."""
-        return SyncEngine._build_local_input(local_path, filename)
-
-    def _build_sync_conflict_entry(
-        self,
-        rom_id: int,
-        filename: str,
-        server: dict,
-        local_path: str | None,
-        local_hash: str | None,
-    ) -> dict:
-        """Delegate to :class:`SyncEngine` — see its docstring."""
-        return self._sync_engine._build_sync_conflict_entry(rom_id, filename, server, local_path, local_hash)
-
-    def _adopt_baseline_hash(self, rom_id_str: str, filename: str, local_hash: str) -> None:
-        """Delegate to :class:`SyncEngine` — see its docstring."""
-        self._sync_engine._adopt_baseline_hash(rom_id_str, filename, local_hash)
-
-    def _sync_rom_saves(self, rom_id: int) -> tuple[int, list[str], list[SaveConflict | dict]]:
-        """Delegate to :class:`SyncEngine` — see its docstring."""
-        return self._sync_engine._sync_rom_saves(rom_id)
-
-    def _is_save_sync_enabled(self) -> bool:
-        """Check if save sync feature is enabled."""
-        return self._save_sync_state.get("settings", {}).get("save_sync_enabled", False)
-
-    def _get_save_status_io(self, rom_id: int, server_saves: list[dict]) -> dict:
-        """Delegate to :class:`StatusService` — see its docstring."""
-        return self._status._get_save_status_io(rom_id, server_saves)
-
-    # ------------------------------------------------------------------
-    # Public async API (callable endpoints)
-    # ------------------------------------------------------------------
-
-    async def ensure_device_registered(self) -> dict:
-        """Ensure this device is registered with the RomM server for save sync tracking."""
-        if not self._is_save_sync_enabled():
-            return {"success": False, "device_id": "", "device_name": "", "disabled": True}
-
-        # Already registered
-        has_device_id = self._save_sync_state.get("device_id")
-        has_server_id = self._save_sync_state.get("server_device_id")
-        if has_device_id and has_server_id:
-            with contextlib.suppress(Exception):
-                await self._loop.run_in_executor(
-                    None,
-                    lambda: self._romm_api.update_device(has_server_id, client_version=self._plugin_version),
-                )
-            return {
-                "success": True,
-                "device_id": self._save_sync_state["device_id"],
-                "device_name": self._save_sync_state.get("device_name", ""),
-                "server_device_id": has_server_id,
-            }
-
-        hostname = socket.gethostname()
-
-        try:
-            result = await self._loop.run_in_executor(
-                None,
-                lambda: self._romm_api.register_device(
-                    name=hostname,
-                    platform="linux",
-                    client="decky-romm-sync",
-                    client_version=self._plugin_version,
-                ),
-            )
-            server_device_id = result.get("id") or result.get("device_id")
-            if server_device_id:
-                self._save_sync_state["device_id"] = str(server_device_id)
-                self._save_sync_state["device_name"] = hostname
-                self._save_sync_state["server_device_id"] = str(server_device_id)
-                self.save_state()
-                self._logger.info(f"Device registered with server: {server_device_id} ({hostname})")
-                return {
-                    "success": True,
-                    "device_id": str(server_device_id),
-                    "device_name": hostname,
-                    "server_device_id": str(server_device_id),
-                }
-        except Exception as e:
-            self._logger.warning(f"Server device registration failed: {e}")
-
-        return {"success": False, "device_id": "", "device_name": "", "error": "registration_failed"}
-
-    async def get_save_status(self, rom_id: int) -> dict:
-        """Get save sync status for a ROM (local files, server saves, conflict state)."""
-        rom_id = int(rom_id)
-
-        server_saves: list[dict] = []
-        try:
-            device_id = self._get_server_device_id()
-            server_saves = await self._loop.run_in_executor(
-                None,
-                lambda: self._retry.with_retry(lambda: self._romm_api.list_saves(rom_id, device_id=device_id)),
-            )
-        except Exception as e:
-            self._log_debug(f"Failed to fetch saves for rom {rom_id}: {e}")
-
-        return await self._loop.run_in_executor(None, self._status._get_save_status_io, rom_id, server_saves)
-
-    async def check_save_status_background(self, rom_id: int) -> None:
-        """Run full save status check in background and emit result to frontend."""
-        try:
-            result = await self.get_save_status(rom_id)
-            if self._emit is not None:
-                await self._emit("save_status_updated", result)
-        except Exception as e:
-            self._log_debug(f"Background save status check failed for rom {rom_id}: {e}")
-
-    async def list_devices(self) -> dict:
-        """List all devices registered with the RomM server for this user."""
-        if not self._is_save_sync_enabled():
-            return {"success": False, "devices": [], "disabled": True}
-        try:
-            own_id = self._get_server_device_id()
-            devices = await self._loop.run_in_executor(
-                None,
-                lambda: self._retry.with_retry(lambda: self._romm_api.list_devices()),
-            )
-            own_id_str = str(own_id or "")
-            enriched = [
-                {**d, "is_current_device": bool(own_id_str) and (str(d.get("id") or "")) == own_id_str} for d in devices
-            ]
-            return {"success": True, "devices": enriched}
-        except Exception as e:
-            self._log_debug(f"list_devices failed: {e}")
-            return {"success": False, "devices": [], "error": "list_failed"}
-
-    def check_core_change(self, rom_id: int) -> dict:
-        """Check if emulator core changed since last sync for a ROM."""
-        if not self._is_save_sync_enabled():
-            return {"changed": False}
-
-        rom_id_str = str(rom_id)
-        save_entry = self._save_sync_state.get("saves", {}).get(rom_id_str)
-        if not save_entry:
-            return {"changed": False}  # Never synced
-
-        stored_core = save_entry.get("last_synced_core")
-        system = save_entry.get("system")
-        if not stored_core or not system:
-            return {"changed": False}
-
-        # Resolve ROM filename for per-game core detection
-        rom_filename = None
-        installed = self._state.get("installed_roms", {}).get(rom_id_str)
-        if installed:
-            file_path = installed.get("file_path", "")
-            if file_path:
-                rom_filename = os.path.basename(file_path)
-
-        # TODO: Core labels come from ES-DE config which may differ from RetroArch's
-        # corename (e.g. "Snes9x - Current" vs "Snes9x"). Align with RetroArch
-        # core names when #208 is resolved.
-        try:
-            active_core, active_label = self._get_active_core(system, rom_filename)
-        except Exception:
-            return {"changed": False}
-
-        changed = detect_core_change(stored_core, active_core)
-
-        if not changed:
-            return {"changed": False}
-
-        # Strip _libretro suffix for display (stored_core is guaranteed non-None here)
-        old_label = stored_core.replace("_libretro", "")
-
-        return {
-            "changed": True,
-            "old_core": stored_core,
-            "new_core": active_core,
-            "old_label": old_label,
-            "new_label": active_label or (active_core.replace("_libretro", "") if active_core else None),
-        }
-
-    async def _refresh_save_sort_state(self, where: str) -> None:
-        """Refresh save-sort state from the live RetroArch config.
-
-        Save-sync must observe fresh save-sort state before computing
-        ``saves_dir``. This call ensures ``detect_save_sort_change`` has
-        run at least once before we read state, closing the race where
-        another frontend detect trigger arrives after our backend entry
-        point. Without this, a direct-Steam-launch with no pre-detect
-        would silently download stale server content to the wrong
-        layout and destroy real user progress during the subsequent
-        migration (#238).
-
-        Graceful degradation: if detect fails (e.g. retroarch.cfg is
-        temporarily unreadable) we log and continue with the
-        previously-known state — save-sync must not abort because of a
-        config read error.
-        """
-        if self._detect_sort_change is None:
-            return
-        try:
-            await self._loop.run_in_executor(None, self._detect_sort_change)
-        except Exception as e:
-            self._logger.warning(
-                "%s: detect_sort_change failed (%s) — proceeding with stale state",
-                where,
-                e,
-            )
-
-    async def pre_launch_sync(self, rom_id: int) -> dict:
-        """Download newer saves from server before game launch."""
-        rom_id = int(rom_id)
-        async with self._rom_lock(rom_id):
-            if not self._is_save_sync_enabled():
-                return {"success": True, "message": "Save sync disabled", "synced": 0}
-
-            # Defense in depth: block pre_launch_sync if a future caller bypasses
-            # the @migration_blocked decorator at the public callable. saves_dir
-            # would otherwise resolve under the new home and silently desync from
-            # files still living at the old home. Internal _sync_rom_saves callers
-            # (sync_all_saves, rollback_to_version) are protected by the decorator
-            # on their own public callables — this guard is for pre_launch_sync.
-            if self._is_retrodeck_migration_pending and self._is_retrodeck_migration_pending():
-                return {
-                    "success": False,
-                    "message": "Pending RetroDECK migration. Open the plugin QAM to migrate or dismiss.",
-                    "synced": 0,
-                    "blocked_by_migration": True,
-                }
-
-            # Refresh save-sort state before the migration gate — see #238.
-            await self._refresh_save_sort_state("pre_launch_sync")
-
-            if self._is_save_sort_changed():
-                return {
-                    "success": False,
-                    "message": "RetroArch save sorting changed — migrate saves in Settings first",
-                    "synced": 0,
-                    "save_sort_changed": True,
-                }
-
-            settings = self._save_sync_state.get("settings", {})
-            if not settings.get("sync_before_launch", True):
-                return {"success": True, "message": "Pre-launch sync disabled", "synced": 0}
-
-            if not self._save_sync_state.get("device_id"):
-                reg = await self.ensure_device_registered()
-                if not reg.get("success"):
-                    return {"success": False, "message": _DEVICE_NOT_REGISTERED}
-
-            synced, errors, conflicts = await self._loop.run_in_executor(None, self._sync_rom_saves, rom_id)
-            self.save_state()
-
-            msg = f"Downloaded {synced} save(s)"
-            if errors:
-                msg += f", {len(errors)} error(s)"
-            return {
-                "success": len(errors) == 0,
-                "message": msg,
-                "synced": synced,
-                "errors": errors,
-                "conflicts": [c if isinstance(c, dict) else asdict(c) for c in conflicts],
-            }
-
-    async def post_exit_sync(self, rom_id: int) -> dict:
-        """Upload changed saves after game exit."""
-        self._logger.info("post_exit_sync called for rom_id=%d", rom_id)
-        rom_id = int(rom_id)
-
-        async with self._rom_lock(rom_id):
-            if not self._is_save_sync_enabled():
-                self._logger.info("post_exit_sync skipped: save sync disabled")
-                return {"success": True, "message": "Save sync disabled", "synced": 0}
-
-            # Defense in depth: same rationale as pre_launch_sync — internal
-            # _sync_rom_saves callers are protected by @migration_blocked on
-            # their public callables; this guard covers post_exit_sync only.
-            if self._is_retrodeck_migration_pending and self._is_retrodeck_migration_pending():
-                self._logger.info("post_exit_sync skipped: retrodeck migration pending")
-                return {
-                    "success": False,
-                    "message": "Pending RetroDECK migration. Open the plugin QAM to migrate or dismiss.",
-                    "synced": 0,
-                    "blocked_by_migration": True,
-                }
-
-            settings = self._save_sync_state.get("settings", {})
-            if not settings.get("sync_after_exit", True):
-                self._logger.info("post_exit_sync skipped: sync_after_exit disabled")
-                return {"success": True, "message": "Post-exit sync disabled", "synced": 0}
-
-            # Refresh save-sort state before _sync_rom_saves reads saves_dir — see #238.
-            await self._refresh_save_sort_state("post_exit_sync")
-
-            try:
-                await self._loop.run_in_executor(None, self._romm_api.heartbeat)
-            except Exception:
-                self._logger.info("post_exit_sync skipped: server offline")
-                return {"success": False, "message": "Server offline", "synced": 0, "offline": True}
-
-            if not self._save_sync_state.get("device_id"):
-                reg = await self.ensure_device_registered()
-                if not reg.get("success"):
-                    return {"success": False, "message": _DEVICE_NOT_REGISTERED}
-
-            synced, errors, conflicts = await self._loop.run_in_executor(None, self._sync_rom_saves, rom_id)
-            self.save_state()
-
-            self._logger.info(
-                "post_exit_sync complete for rom_id=%d: synced=%d, errors=%d, conflicts=%d",
-                rom_id,
-                synced,
-                len(errors),
-                len(conflicts),
-            )
-
-            msg = f"Uploaded {synced} save(s)"
-            if errors:
-                msg += f", {len(errors)} error(s)"
-            if conflicts:
-                msg += f", {len(conflicts)} conflict(s)"
-            return {
-                "success": len(errors) == 0,
-                "message": msg,
-                "synced": synced,
-                "errors": errors,
-                "conflicts": [c if isinstance(c, dict) else asdict(c) for c in conflicts],
-            }
-
-    async def sync_rom_saves(self, rom_id: int) -> dict:
-        """Bidirectional sync for a single ROM (manual trigger from game detail)."""
-        rom_id = int(rom_id)
-        async with self._rom_lock(rom_id):
-            if not self._is_save_sync_enabled():
-                return {"success": False, "message": _SYNC_DISABLED_MSG, "synced": 0}
-
-            # Refresh save-sort state before _sync_rom_saves reads saves_dir — see #238.
-            # Manual sync paths must observe fresh sort state too: a user could
-            # edit retroarch.cfg outside of a session and then trigger a manual
-            # sync before any detect has fired.
-            await self._refresh_save_sort_state("sync_rom_saves")
-
-            if not self._save_sync_state.get("device_id"):
-                reg = await self.ensure_device_registered()
-                if not reg.get("success"):
-                    return {"success": False, "message": _DEVICE_NOT_REGISTERED}
-
-            synced, errors, conflicts = await self._loop.run_in_executor(None, self._sync_rom_saves, rom_id)
-            self.save_state()
-
-            msg = f"Synced {synced} save(s)"
-            if errors:
-                msg += f", {len(errors)} error(s)"
-            if conflicts:
-                msg += f", {len(conflicts)} conflict(s)"
-            return {
-                "success": len(errors) == 0,
-                "message": msg,
-                "synced": synced,
-                "errors": errors,
-                "conflicts": [c if isinstance(c, dict) else asdict(c) for c in conflicts],
-            }
-
-    async def get_save_slots(self, rom_id: int) -> dict:
-        """List available save slots for a ROM.
-
-        Merges server slots with locally-created slots. Persists the merged
-        result so local slots survive restarts. Promotes local slots to server
-        when they appear on the server. Removes server slots that no longer
-        exist on the server (unless they are the active_slot).
-        """
-        rom_id = int(rom_id)
-        if not self._is_save_sync_enabled():
-            return {"success": False, "slots": [], "active_slot": "default"}
-
-        rom_id_str = str(rom_id)
-        device_id = self._get_server_device_id()
-        rom_state = self._save_sync_state.get("saves", {}).get(rom_id_str, {})
-        active_slot = rom_state.get(
-            "active_slot",
-            self._save_sync_state.get("settings", {}).get("default_slot", "default"),
-        )
-        persisted_slots: dict[str, dict] = rom_state.get("slots", {})
-
-        # Fetch server slots
-        server_slots_list: list[dict] = []
-        try:
-            summary = await self._loop.run_in_executor(
-                None,
-                lambda: self._retry.with_retry(
-                    lambda: self._romm_api.get_save_summary(rom_id, device_id=device_id),
-                ),
-            )
-            server_slots_list = summary.get("slots", [])
-        except Exception as e:
-            self._log_debug(f"Failed to fetch save slots for rom {rom_id}: {e}")
-
-        # Merge: update persisted slots with server data, promote local→server
-        merged: dict[str, dict] = {}
-        for s in server_slots_list:
-            raw = s.get("slot") or s.get("slot_name")
-            name = raw if raw else ""
-            merged[name] = {
-                "source": "server",
-                "count": s.get("count", 0),
-                "latest_updated_at": (s.get("latest") or {}).get("updated_at"),
-            }
-
-        self._merge_persisted_slots(persisted_slots, merged, active_slot)
-
-        # Persist merged slots in state
-        game_entry = self._save_sync_state.setdefault("saves", {}).setdefault(rom_id_str, {})
-        game_entry["slots"] = merged
-        self.save_state()
-
-        # Build response list
-        result_slots = [
-            {
-                "slot": name,
-                "source": info.get("source", "server"),
-                "count": info.get("count", 0),
-                "latest_updated_at": info.get("latest_updated_at"),
-            }
-            for name, info in sorted(merged.items())
-        ]
-
-        return {"success": True, "slots": result_slots, "active_slot": active_slot}
-
-    @staticmethod
-    def _merge_persisted_slots(
-        persisted: dict[str, dict],
-        merged: dict[str, dict],
-        active_slot: str | None,
-    ) -> None:
-        """Add persisted local slots (or the active slot) that aren't on the server.
-
-        Mutates ``merged`` in place. Local slots are always kept. A persisted
-        server slot that's gone from the server is dropped unless it's the
-        active slot — we want to keep the UI functional until the user
-        explicitly switches away.
-        """
-        for name, info in persisted.items():
-            if name in merged:
-                continue
-            if info.get("source") == "local":
-                merged[name] = {"source": "local", "count": 0, "latest_updated_at": None}
-            elif info.get("source") == "server" and name == (active_slot or ""):
-                merged[name] = {"source": "server", "count": 0, "latest_updated_at": None}
-
-    async def get_slot_saves(self, rom_id: int, slot: str) -> dict:
-        """Fetch server save files for a specific slot.
-
-        Used by the frontend to show save files when expanding an inactive slot panel.
-        Lightweight — no local file scanning or conflict detection.
-        """
-        rom_id = int(rom_id)
-        slot = str(slot).strip() if slot else ""
-
-        if not self._is_save_sync_enabled():
-            return {"success": False, "slot": slot, "saves": [], "error": _SYNC_DISABLED_MSG}
-
-        device_id = self._get_server_device_id()
-
-        try:
-            server_saves: list[dict] = await self._loop.run_in_executor(
-                None,
-                lambda: self._retry.with_retry(
-                    lambda: self._romm_api.list_saves(rom_id, device_id=device_id, slot=slot),
-                ),
-            )
-            saves = [
-                {
-                    "filename": s["file_name"],
-                    "id": s["id"],
-                    "size": s.get("file_size_bytes"),
-                    "updated_at": s.get("updated_at", ""),
-                    "emulator": s.get("emulator", ""),
-                }
-                for s in server_saves
-            ]
-            return {"success": True, "slot": slot, "saves": saves}
-        except Exception as e:
-            return {"success": False, "slot": slot, "saves": [], "error": str(e)}
-
-    def set_game_slot(self, rom_id: int, slot: str) -> dict:
-        """Set the active save slot for a specific game.
-
-        If the slot doesn't exist yet (not on server), it is persisted
-        as a local slot. It will be promoted to server once a save is
-        uploaded to it.
-        """
-        rom_id = int(rom_id)
-        slot_str = str(slot).strip() if slot else ""
-        # Empty string = legacy mode (None slot)
-        resolved_slot: str | None = slot_str if slot_str else None
-
-        rom_id_str = str(rom_id)
-        saves = self._save_sync_state.setdefault("saves", {})
-        if rom_id_str not in saves:
-            saves[rom_id_str] = {"files": {}, "active_slot": resolved_slot}
-        else:
-            saves[rom_id_str]["active_slot"] = resolved_slot
-
-        # Ensure slot is in the persisted slots dict (use "" as key for legacy/None)
-        slot_key = resolved_slot if resolved_slot is not None else ""
-        slots_dict: dict[str, dict] = saves[rom_id_str].setdefault("slots", {})
-        if slot_key not in slots_dict:
-            slots_dict[slot_key] = {"source": "local", "count": 0, "latest_updated_at": None}
-
-        self.save_state()
-        self._loop.create_task(self.check_save_status_background(rom_id))
-        return {"success": True, "active_slot": resolved_slot}
-
-    def _check_slot_switch_readiness(self, rom_id: int) -> dict:
-        """Check whether it is safe to switch slots for this ROM.
-
-        A switch is unsafe if local files have changed since the last sync
-        to the current slot — those changes would be lost.
-        Files that were never synced do not block (they'll be deleted on switch).
-
-        Returns ``{"ready": True}`` or
-        ``{"ready": False, "reason": str, "files": list[str]}``.
-        """
-        rom_id_str = str(rom_id)
-        save_state = self._save_sync_state["saves"].get(rom_id_str, {})
-        files_state = save_state.get("files", {})
-
-        pending: list[str] = []
-        local_files = self._find_save_files(rom_id)
-        for lf in local_files:
-            filename = lf["filename"]
-            file_state = files_state.get(filename, {})
-            last_sync_hash = file_state.get("last_sync_hash")
-            if last_sync_hash:
-                current_hash = self._file_md5(lf["path"])
-                if current_hash != last_sync_hash:
-                    pending.append(filename)
-
-        if pending:
-            return {"ready": False, "reason": "pending_uploads", "files": pending}
-
-        return {"ready": True}
-
-    async def switch_slot(self, rom_id: int, new_slot: str) -> dict:
-        """Switch the active save slot with immediate state sync.
-
-        Pre-checks (all must pass):
-        1. Save sync must be enabled.
-        2. ROM must be installed.
-        3. No local files with pending changes (changed since last sync to current slot).
-        4. Server must be reachable.
-
-        On success:
-        - If the new slot has server saves: downloads them, replacing local files.
-        - If the new slot is empty: deletes local save files (fresh start).
-        - Never uploads — saves are not carried between slots.
-        """
-        rom_id = int(rom_id)
-        rom_id_str = str(rom_id)
-
-        # 1. Save sync must be enabled
-        if not self._is_save_sync_enabled():
-            return {"success": False, "reason": "sync_disabled"}
-
-        # 2. Slot normalisation (empty → None for legacy mode)
-        slot_str = str(new_slot).strip() if new_slot else ""
-        resolved_slot: str | None = slot_str if slot_str else None
-
-        # 3. ROM must be installed
-        info = self._get_rom_save_info(rom_id)
-        if not info:
-            return {"success": False, "reason": "not_installed"}
-
-        saves_dir = info["saves_dir"]
-        system = info["system"]
-
-        # 4. Check for pending local changes (hashing — run in executor)
-        readiness = await self._loop.run_in_executor(None, self._check_slot_switch_readiness, rom_id)
-        if not readiness.get("ready"):
-            return {
-                "success": False,
-                "reason": readiness.get("reason", "pending_uploads"),
-                "files": readiness.get("files", []),
-            }
-
-        # 5. Fetch server saves for the new slot (also proves server is reachable)
-        device_id = self._get_server_device_id()
-        try:
-            all_server_saves: list[dict] = await self._loop.run_in_executor(
-                None,
-                lambda: self._retry.with_retry(
-                    lambda: self._romm_api.list_saves(rom_id, device_id=device_id),
-                ),
-            )
-        except Exception:
-            return {"success": False, "reason": "server_unreachable"}
-
-        # Filter to the target slot (FakeSaveApi doesn't filter, real API may not either)
-        # Normalize "" and None both to None before comparing (legacy saves may use either)
-        slot_saves = [s for s in all_server_saves if (s.get("slot") or None) == resolved_slot]
-
-        # 6. Update active slot in state
-        self.set_game_slot(rom_id, new_slot)
-
-        # 7. Sync local state to match the new slot
-        if slot_saves:
-            # New slot has server saves — download them, replacing local files.
-            # rom_name is guaranteed by the earlier ``info`` check (line 1642).
-            await self._loop.run_in_executor(
-                None,
-                self._do_switch_downloads,
-                slot_saves,
-                saves_dir,
-                rom_id_str,
-                system,
-                info["rom_name"],
-            )
-        else:
-            # New slot is empty — delete local save files for a fresh start
-            await self._loop.run_in_executor(
-                None,
-                self._delete_local_saves_for_switch,
-                rom_id,
-                rom_id_str,
-            )
-
-        # 8. Update last_sync_check_at
-        save_entry = self._save_sync_state["saves"].setdefault(rom_id_str, {})
-        save_entry["last_sync_check_at"] = datetime.now(UTC).isoformat()
-        self.save_state()
-
-        # 9. Return fresh status
-        save_status = await self.get_save_status(rom_id)
-        return {"success": True, "save_status": save_status}
-
-    def _do_switch_downloads(
-        self,
-        slot_saves: list[dict],
-        saves_dir: str,
-        rom_id_str: str,
-        system: str,
-        rom_name: str,
-    ) -> None:
-        """Download all saves from *slot_saves* into *saves_dir*.
-
-        Each save lands at ``<saves_dir>/<rom_name>.<server.file_extension>`` —
-        the canonical RetroArch path. Runs synchronously; call via
-        ``run_in_executor``.
-        """
-        for server_save in slot_saves:
-            target = _local_save_target(server_save, rom_name)
-            self._do_download_save(server_save, saves_dir, target, rom_id_str, system)
-
-    def _delete_local_saves_for_switch(self, rom_id: int, rom_id_str: str) -> None:
-        """Delete local save files and clear file tracking state for a slot switch.
-
-        Unlike delete_local_saves (the callable), this preserves slot config
-        (active_slot, slot_confirmed, slots dict) and only clears files + tracking.
-        Runs synchronously — call via run_in_executor.
-        """
-        local_files = self._find_save_files(rom_id)
-        for lf in local_files:
-            try:
-                os.remove(lf["path"])
-                self._log_debug(f"Deleted local save for switch: {lf['filename']}")
-            except Exception as e:
-                self._log_debug(f"Failed to delete {lf['filename']} during switch: {e}")
-
-        # Clear file tracking state (but keep slot config)
-        save_entry = self._save_sync_state.get("saves", {}).get(rom_id_str, {})
-        save_entry["files"] = {}
-
-    # ------------------------------------------------------------------
-    # Save Setup Wizard
-    # ------------------------------------------------------------------
-
-    def is_save_tracking_configured(self, rom_id: int) -> dict:
-        """Check if save slot tracking is configured for a game.
-
-        Fast, synchronous check — reads only from local state.
-        Returns {"configured": bool, "active_slot": str|None}
-        """
-        rom_id_str = str(int(rom_id))
-        game_state = self._save_sync_state["saves"].get(rom_id_str, {})
-        configured = game_state.get("slot_confirmed", False)
-        active_slot = game_state.get("active_slot") if configured else None
-        return {"configured": configured, "active_slot": active_slot}
-
-    async def get_save_setup_info(self, rom_id: int) -> dict:
-        """Get info needed for the first-sync setup wizard.
-
-        Fetches server saves, checks local files, determines which
-        scenario (A-E) applies so the frontend can display the right UI.
-        """
-        rom_id = int(rom_id)
-        rom_id_str = str(rom_id)
-
-        # Local saves
-        local_files = self._find_save_files(rom_id)
-        local_file_info = []
-        for lf in local_files:
-            local_file_info.append(
-                {
-                    "filename": lf["filename"],
-                    "size": os.path.getsize(lf["path"]) if os.path.isfile(lf["path"]) else 0,
-                }
-            )
-
-        # Server saves
-        server_saves: list[dict] = []
-        device_id = self._get_server_device_id()
-        try:
-            server_saves = await self._loop.run_in_executor(
-                None,
-                lambda: self._retry.with_retry(
-                    lambda: self._romm_api.list_saves(rom_id, device_id=device_id),
-                ),
-            )
-        except Exception as e:
-            self._log_debug(f"get_save_setup_info({rom_id}): failed to list saves: {e}")
-
-        # Group server saves by slot
-        slots_map: dict[str | None, list[dict]] = {}
-        for ss in server_saves:
-            slot_key = ss.get("slot")
-            slots_map.setdefault(slot_key, []).append(ss)
-
-        server_slots = []
-        for slot_key, saves in slots_map.items():
-            latest = max((s.get("updated_at", "") for s in saves), default=None)
-            server_slots.append(
-                {
-                    "slot": slot_key,
-                    "saves": [
-                        {
-                            "id": s.get("id"),
-                            "file_name": s.get("file_name", ""),
-                            "emulator": s.get("emulator", ""),
-                            "updated_at": s.get("updated_at", ""),
-                            "file_size_bytes": s.get("file_size_bytes", 0),
-                        }
-                        for s in saves
-                    ],
-                    "count": len(saves),
-                    "latest_updated_at": latest,
-                }
-            )
-
-        # State info
-        game_state = self._save_sync_state["saves"].get(rom_id_str, {})
-        default_slot = self._save_sync_state.get("settings", {}).get("default_slot", "default")
-        slot_confirmed = game_state.get("slot_confirmed", False)
-        active_slot = game_state.get("active_slot") if slot_confirmed else None
-
-        return {
-            "has_local_saves": len(local_files) > 0,
-            "local_files": local_file_info,
-            "server_slots": server_slots,
-            "default_slot": default_slot,
-            "slot_confirmed": slot_confirmed,
-            "active_slot": active_slot,
-        }
-
-    async def confirm_slot_choice(
-        self,
-        rom_id: int,
-        chosen_slot: str,
-        migrate_from_slot: str | None | object = _NO_MIGRATION,
-    ) -> dict:
-        """Confirm which slot to use for a game's save sync.
-
-        Sets slot_confirmed=true and active_slot in state.
-
-        If migrate_from_slot is provided (can be None for legacy no-slot saves),
-        migrates saves: upload local files to chosen_slot, then delete old server saves.
-        Pass _NO_MIGRATION sentinel (the default) to skip migration.
-        """
-        rom_id = int(rom_id)
-        rom_id_str = str(rom_id)
-        chosen_slot = str(chosen_slot).strip()
-        if not chosen_slot:
-            return {"success": False, "needs_conflict_resolution": False, "message": "Slot name cannot be empty"}
-
-        # Update state
-        saves = self._save_sync_state.setdefault("saves", {})
-        if rom_id_str not in saves:
-            saves[rom_id_str] = {"files": {}}
-        saves[rom_id_str]["active_slot"] = chosen_slot
-        saves[rom_id_str]["slot_confirmed"] = True
-
-        # Migration: re-upload local files to new slot, delete old server saves
-        if migrate_from_slot is not _NO_MIGRATION:
-            # migrate_from_slot can be None (legacy no-slot) or a string slot name
-            from_slot: str | None = migrate_from_slot if isinstance(migrate_from_slot, str) else None
-            try:
-                await self._migrate_slot_saves(rom_id, rom_id_str, chosen_slot, from_slot)
-            except Exception as e:
-                self._logger.warning(f"confirm_slot_choice({rom_id}): migration failed: {e}")
-                self.save_state()
-                return {
-                    "success": True,
-                    "needs_conflict_resolution": False,
-                    "message": f"Slot confirmed but migration failed: {e}",
-                }
-
-        self.save_state()
-        return {"success": True, "needs_conflict_resolution": False, "message": "Slot confirmed"}
-
-    async def _migrate_slot_saves(
-        self,
-        rom_id: int,
-        rom_id_str: str,
-        chosen_slot: str,
-        migrate_from_slot: str | None,
-    ) -> None:
-        """Migrate server saves from one slot to another.
-
-        For each local file: upload with new slot, then delete old server save.
-        Safe order: POST first, DELETE after.
-        """
-        device_id = self._get_server_device_id()
-
-        # Find server saves in the old slot
-        all_saves = await self._loop.run_in_executor(
-            None,
-            lambda: self._retry.with_retry(
-                lambda: self._romm_api.list_saves(rom_id, device_id=device_id),
-            ),
-        )
-        old_slot_saves = [s for s in all_saves if s.get("slot") == migrate_from_slot]
-        if not old_slot_saves:
-            return
-
-        # Get local files for re-upload
-        local_files = self._find_save_files(rom_id)
-        local_by_name = {lf["filename"]: lf for lf in local_files}
-
-        # Resolve emulator tag
-        info = self._get_rom_save_info(rom_id)
-        system = info["system"] if info else ""
-        installed = self._state["installed_roms"].get(rom_id_str, {})
-        rom_filename = os.path.basename(installed.get("file_path", "")) or None
-        core_so, _label = self._get_active_core(system, rom_filename)
-        emulator = build_emulator_tag(core_so)
-
-        ids_to_delete: list[int] = []
-
-        for old_save in old_slot_saves:
-            fname = old_save.get("file_name", "")
-            local_file = local_by_name.get(fname)
-            if local_file and os.path.isfile(local_file["path"]):
-                # Upload to new slot
-                await self._loop.run_in_executor(
-                    None,
-                    lambda lf=local_file, em=emulator: self._retry.with_retry(
-                        lambda: self._romm_api.upload_save(
-                            rom_id,
-                            lf["path"],
-                            em,
-                            device_id=device_id,
-                            slot=chosen_slot,
-                        ),
-                    ),
-                )
-            old_id = old_save.get("id")
-            if old_id is not None:
-                ids_to_delete.append(old_id)
-
-        # Delete old saves
-        if ids_to_delete:
-            await self._loop.run_in_executor(
-                None,
-                lambda: self._retry.with_retry(
-                    lambda: self._romm_api.delete_server_saves(ids_to_delete),
-                ),
-            )
-
-    async def sync_all_saves(self) -> dict:
-        """Manual full sync of all ROMs with shortcuts (both directions)."""
-        if not self._is_save_sync_enabled():
-            return {"success": False, "message": _SYNC_DISABLED_MSG, "synced": 0, "conflicts": 0}
-
-        # Refresh save-sort state before _sync_rom_saves reads saves_dir — see #238.
-        # Manual sync paths must observe fresh sort state too: a user could
-        # edit retroarch.cfg outside of a session and then trigger a manual
-        # sync before any detect has fired.
-        await self._refresh_save_sort_state("sync_all_saves")
-
-        if not self._save_sync_state.get("device_id"):
-            reg = await self.ensure_device_registered()
-            if not reg.get("success"):
-                return {"success": False, "message": _DEVICE_NOT_REGISTERED}
-
-        total_synced = 0
-        total_errors: list[str] = []
-        all_conflicts: list[SaveConflict | dict] = []
-        rom_count = 0
-
-        # Only iterate installed ROMs — non-installed ROMs have no save files
-        rom_ids = set(self._state["installed_roms"].keys())
-        self._log_debug(f"sync_all_saves: {len(rom_ids)} ROMs to check")
-
-        for rom_id_str in sorted(rom_ids):
-            rom_count += 1
-            rom_id_int = int(rom_id_str)
-            async with self._rom_lock(rom_id_int):
-                synced, errors, conflicts = await self._loop.run_in_executor(None, self._sync_rom_saves, rom_id_int)
-            total_synced += synced
-            total_errors.extend(errors)
-            all_conflicts.extend(conflicts)
-
-        self.save_state()
-
-        conflicts_count = len(all_conflicts)
-        msg = f"Synced {total_synced} save(s) across {rom_count} ROM(s)"
-        if total_errors:
-            msg += f", {len(total_errors)} error(s)"
-        if conflicts_count:
-            msg += f", {conflicts_count} conflict(s)"
-        return {
-            "success": len(total_errors) == 0,
-            "message": msg,
-            "synced": total_synced,
-            "conflicts": conflicts_count,
-            "conflicts_list": [c if isinstance(c, dict) else asdict(c) for c in all_conflicts],
-            "roms_checked": rom_count,
-            "errors": total_errors,
-        }
-
-    async def resolve_sync_conflict(
-        self,
-        rom_id: int,
-        filename: str,
-        action: str,
-    ) -> dict:
-        """Resolve a pending sync conflict (true two-sided divergence).
-
-        Reached when ``compute_sync_action`` returned ``Conflict`` — the
-        server moved AND local diverged from baseline, so the user picked a
-        side via the conflict UI.
-
-        ``action`` is one of:
-
-        - ``"keep_local"`` — push local to the current server save (PUT). When
-          the local content already matches the server's content hash we adopt
-          it silently without re-uploading.
-        - ``"use_server"`` — download the current server save, replacing local.
-        """
-        rom_id = int(rom_id)
-        rom_id_str = str(rom_id)
-
-        if action not in ("keep_local", "use_server"):
-            return {"success": False, "message": f"Invalid action: {action}"}
-
-        async with self._rom_lock(rom_id):
-            info = self._get_rom_save_info(rom_id)
-            if not info:
-                return {"success": False, "message": "ROM not installed"}
-            system = info["system"]
-            saves_dir = info["saves_dir"]
-
-            try:
-                device_id = self._get_server_device_id()
-                server_saves = await self._loop.run_in_executor(
-                    None,
-                    lambda: self._retry.with_retry(
-                        lambda: self._romm_api.list_saves(rom_id, device_id=device_id),
-                    ),
-                )
-            except Exception as e:
-                _code, _msg = classify_error(e)
-                return {"success": False, "message": f"Failed to fetch saves: {_msg}"}
-
-            save_state = self._save_sync_state["saves"].get(rom_id_str, {})
-            active_slot = save_state.get("active_slot")
-            server_in_slot = self._filter_server_saves_to_slot(server_saves, active_slot)
-            if not server_in_slot:
-                return {"success": False, "message": "No server save in active slot"}
-            server = max(server_in_slot, key=lambda s: parse_iso_to_epoch(s.get("updated_at")) or 0.0)
-
-            try:
-                if action == "use_server":
-                    await self._loop.run_in_executor(
-                        None,
-                        self._apply_resolve_use_server,
-                        rom_id_str,
-                        server,
-                        saves_dir,
-                        system,
-                        info["rom_name"],
-                    )
-                    self._logger.info(
-                        "resolve_sync_conflict(rom_id=%d, filename=%s, action=%s) -> success",
-                        rom_id,
-                        filename,
-                        action,
-                    )
-                    return {"success": True, "action": "use_server"}
-
-                # keep_local
-                await self._loop.run_in_executor(
-                    None,
-                    self._apply_resolve_keep_local,
-                    rom_id,
-                    rom_id_str,
-                    filename,
-                    server,
-                    saves_dir,
-                    system,
-                )
-                self._logger.info(
-                    "resolve_sync_conflict(rom_id=%d, filename=%s, action=%s) -> success",
-                    rom_id,
-                    filename,
-                    action,
-                )
-                return {"success": True, "action": "keep_local"}
-            except Exception as e:
-                self._logger.error(f"resolve_sync_conflict({rom_id}, {filename}, {action}) failed: {e}")
-                return {"success": False, "message": str(e)}
-
-    def _apply_resolve_use_server(
-        self,
-        rom_id_str: str,
-        server: dict,
-        saves_dir: str,
-        system: str,
-        rom_name: str,
-    ) -> None:
-        """Download *server* into the canonical local save file and update state.
-
-        The write path is always ``<rom_name>.<server.file_extension>`` — the
-        path RetroArch reads. Drives state-key consistency too:
-        ``_update_file_sync_state`` receives the same target name the file
-        lands at.
-        """
-        target = _local_save_target(server, rom_name)
-        self._do_download_save(server, saves_dir, target, rom_id_str, system)
-        self.save_state()
-
-    def _apply_resolve_keep_local(
-        self,
-        rom_id: int,
-        rom_id_str: str,
-        filename: str,
-        server: dict,
-        saves_dir: str,
-        system: str,
-    ) -> None:
-        """Push the local file to *server* (PUT). Adopt-without-upload when the
-        local content already matches the server's content hash.
-        """
-        local_path = os.path.join(saves_dir, filename)
-        if not os.path.isfile(local_path):
-            raise FileNotFoundError(f"Local save not found: {local_path}")
-        local_hash = self._file_md5(local_path)
-        try:
-            server_hash = self._retry.with_retry(self._get_server_save_hash, server)
-        except Exception:
-            server_hash = None
-
-        if server_hash and local_hash == server_hash:
-            # Hashes match — adopt server's id without re-uploading.
-            self._log_debug(
-                f"keep_local: hash matches server, adopting without upload (rom={rom_id} filename={filename})"
-            )
-            saves = self._save_sync_state.setdefault("saves", {})
-            rom_entry = saves.setdefault(rom_id_str, {"files": {}})
-            files = rom_entry.setdefault("files", {})
-            file_state = files.setdefault(filename, {})
-            file_state["tracked_save_id"] = server.get("id")
-            file_state["last_sync_hash"] = local_hash
-            file_state["last_sync_at"] = datetime.now(UTC).isoformat()
-            file_state["last_sync_server_updated_at"] = server.get("updated_at", "")
-            file_state["last_sync_server_size"] = server.get("file_size_bytes")
-            file_state["last_sync_local_mtime"] = os.path.getmtime(local_path)
-            file_state["last_sync_local_size"] = os.path.getsize(local_path)
-            self.save_state()
-            return
-
-        # Upload local content as a PUT against the existing server save.
-        self._do_upload_save(rom_id, local_path, filename, rom_id_str, system, server)
-        self.save_state()
-
-    # ------------------------------------------------------------------
-    # Version History API
-    # ------------------------------------------------------------------
-
-    async def list_file_versions(self, rom_id: int, slot: str, filename: str) -> list[dict]:
-        """Delegate to :class:`VersionsService` — see its docstring."""
-        return await self._versions.list_file_versions(rom_id, slot, filename)
-
-    async def rollback_to_version(self, rom_id: int, slot: str, save_id: int) -> dict:
-        """Delegate to :class:`VersionsService` — see its docstring."""
-        return await self._versions.rollback_to_version(rom_id, slot, save_id)
-
-    def get_save_sync_settings(self) -> dict:
-        """Return current save sync settings."""
-        settings = self._save_sync_state.get("settings", {})
-        # Defensive defaults for keys added after initial release
-        settings.setdefault("default_slot", "default")
-        settings.setdefault("autocleanup_limit", 10)
-        if not self._save_sync_state.get("settings"):
-            settings.setdefault("save_sync_enabled", False)
-            settings.setdefault("sync_before_launch", True)
-            settings.setdefault("sync_after_exit", True)
-        return settings
-
-    @staticmethod
-    def _sanitize_setting(key: str, value: object) -> tuple[object, bool]:
-        """Validate and coerce a single settings key/value pair.
-
-        Returns (coerced_value, skip) where skip=True means the value should
-        be discarded (e.g. empty slot name).
-        """
-        if key == "default_slot":
-            if value is None:
-                return None, False  # None = legacy mode
-            coerced = str(value).strip()
-            return (coerced if coerced else None), False  # empty -> None
-        if key == "autocleanup_limit":
-            return max(1, int(value)), False  # type: ignore[arg-type]
-        if key in ("save_sync_enabled", "sync_before_launch", "sync_after_exit"):
-            return bool(value), False
-        return value, False
-
-    def update_save_sync_settings(self, settings: dict) -> dict:
-        """Update save sync settings (sync toggles, slot, etc.)."""
-        allowed_keys = {
-            "save_sync_enabled",
-            "sync_before_launch",
-            "sync_after_exit",
-            "default_slot",
-            "autocleanup_limit",
-        }
-
-        current = self._save_sync_state.setdefault("settings", {})
-
-        for key, value in settings.items():
-            if key not in allowed_keys:
-                continue
-            value, skip = self._sanitize_setting(key, value)
-            if skip:
-                continue
-            current[key] = value
-
-        self.save_state()
-        return {"success": True, "settings": current}
-
-    def delete_local_saves(self, rom_id: int) -> dict:
-        """Delete local save files (.srm, .rtc) for a ROM."""
-        rom_id = int(rom_id)
-        rom_id_str = str(rom_id)
-
-        files = self._find_save_files(rom_id)
-        if not files:
-            return {"success": True, "deleted_count": 0, "message": "No local save files found"}
-
-        deleted = 0
-        errors = []
-        for f in files:
-            try:
-                os.remove(f["path"])
-                deleted += 1
-            except Exception as e:
-                errors.append(f"{f['filename']}: {e}")
-
-        # Clean up sync state for this ROM
-        self._save_sync_state.get("saves", {}).pop(rom_id_str, None)
-        self.save_state()
-
-        if errors:
-            return {
-                "success": False,
-                "deleted_count": deleted,
-                "message": f"Deleted {deleted} file(s), {len(errors)} error(s)",
-            }
-        return {
-            "success": True,
-            "deleted_count": deleted,
-            "message": f"Deleted {deleted} save file(s)",
-        }
-
-    def delete_platform_saves(self, platform_slug: str) -> dict:
-        """Delete local save files for all installed ROMs on a platform."""
-        total_deleted = 0
-        total_errors: list[str] = []
-        rom_count = 0
-
-        for rom_id_str, entry in self._state["installed_roms"].items():
-            if entry.get("platform_slug") != platform_slug:
-                continue
-            rom_count += 1
-            rom_id = int(rom_id_str)
-            files = self._find_save_files(rom_id)
-            for f in files:
-                try:
-                    os.remove(f["path"])
-                    total_deleted += 1
-                except Exception as e:
-                    total_errors.append(f"{f['filename']}: {e}")
-            # Clean up sync state
-            self._save_sync_state.get("saves", {}).pop(rom_id_str, None)
-
-        self.save_state()
-
-        if total_errors:
-            return {
-                "success": False,
-                "deleted_count": total_deleted,
-                "message": (f"Deleted {total_deleted} file(s) from {rom_count} ROM(s), {len(total_errors)} error(s)"),
-            }
-        return {
-            "success": True,
-            "deleted_count": total_deleted,
-            "message": f"Deleted {total_deleted} save file(s) from {rom_count} ROM(s)",
-        }
-
-    # ------------------------------------------------------------------
-    # Slot deletion
-    # ------------------------------------------------------------------
-
-    def _validate_slot_operation(self, rom_id: int, slot: str) -> dict | tuple[str, dict, dict[str, dict]]:
-        """Shared validation for slot delete operations.
-
-        Returns an error dict on failure, or a (rom_id_str, save_state, slots_dict)
-        tuple on success.
-        """
-        if not self._is_save_sync_enabled():
-            return {"success": False, "reason": "disabled"}
-        if not self._get_rom_save_info(rom_id):
-            return {"success": False, "reason": "not_installed"}
-        rom_id_str = str(rom_id)
-        save_state = self._save_sync_state.get("saves", {}).get(rom_id_str, {})
-        slots_dict: dict[str, dict] = save_state.get("slots", {})
-        if slot not in slots_dict:
-            return {"success": False, "reason": "not_found"}
-        return rom_id_str, save_state, slots_dict
-
-    async def get_slot_delete_info(self, rom_id: int, slot: str) -> dict:
-        """Return info about what deleting a slot would do, for the confirmation modal."""
-        rom_id = int(rom_id)
-        slot = str(slot).strip() if slot else ""
-
-        result = self._validate_slot_operation(rom_id, slot)
-        if isinstance(result, dict):
-            return result
-        _rom_id_str, save_state, slots_dict = result
-
-        slot_info = slots_dict[slot]
-        source = slot_info.get("source", "server")
-        active_slot = save_state.get("active_slot")
-        is_active = slot == (active_slot or "")
-
-        # Server save count
-        server_save_ids: list[int] = []
-        if source == "server":
-            device_id = self._get_server_device_id()
-            try:
-                server_saves: list[dict] = await self._loop.run_in_executor(
-                    None,
-                    lambda: self._retry.with_retry(
-                        lambda: self._romm_api.list_saves(rom_id, device_id=device_id, slot=slot),
-                    ),
-                )
-                server_save_ids = [s["id"] for s in server_saves]
-            except Exception as e:
-                self._log_debug(f"get_slot_delete_info: failed to list saves for slot '{slot}': {e}")
-
-        # Local tracked files pointing to server saves in this slot
-        files_state = save_state.get("files", {})
-        local_filenames: list[str] = []
-        if server_save_ids:
-            id_set = set(server_save_ids)
-            for filename, fstate in files_state.items():
-                if fstate.get("tracked_save_id") in id_set:
-                    local_filenames.append(filename)
-
-        return {
-            "success": True,
-            "slot": slot,
-            "source": source,
-            "server_save_count": len(server_save_ids),
-            "server_save_ids": server_save_ids,
-            "local_file_count": len(local_filenames),
-            "local_filenames": local_filenames,
-            "is_active": is_active,
-        }
-
-    async def _delete_server_slot_saves(self, rom_id: int, slot: str) -> dict:
-        """Delete all server saves in a slot. Returns result dict with count and IDs."""
-        device_id = self._get_server_device_id()
-        try:
-            server_saves: list[dict] = await self._loop.run_in_executor(
-                None,
-                lambda: self._retry.with_retry(
-                    lambda: self._romm_api.list_saves(rom_id, device_id=device_id, slot=slot),
-                ),
-            )
-            save_ids = [s["id"] for s in server_saves]
-            if save_ids:
-                await self._loop.run_in_executor(
-                    None,
-                    lambda: self._retry.with_retry(
-                        lambda: self._romm_api.delete_server_saves(save_ids),
-                    ),
-                )
-            return {"success": True, "count": len(save_ids), "ids": set(save_ids)}
-        except Exception as e:
-            self._logger.warning(f"delete_slot: server delete failed for slot '{slot}': {e}")
-            return {
-                "success": False,
-                "reason": "server_error",
-                "message": f"Failed to delete server saves: {e}",
-            }
-
-    async def delete_slot(self, rom_id: int, slot: str) -> dict:
-        """Delete a save slot and all its saves (local state + server if applicable)."""
-        rom_id = int(rom_id)
-        slot = str(slot).strip() if slot else ""
-
-        result = self._validate_slot_operation(rom_id, slot)
-        if isinstance(result, dict):
-            return result
-        _rom_id_str, save_state, slots_dict = result
-
-        effective_active = save_state.get("active_slot") or ""
-        if slot == effective_active:
-            return {
-                "success": False,
-                "reason": "active_slot",
-                "message": "Cannot delete the active slot. Switch to a different slot first.",
-            }
-
-        slot_info = slots_dict[slot]
-        source = slot_info.get("source", "server")
-
-        deleted_server_saves = 0
-        cleaned_files = 0
-        deleted_ids: set[int] = set()
-
-        if source == "server":
-            result = await self._delete_server_slot_saves(rom_id, slot)
-            if not result["success"]:
-                return result
-            deleted_server_saves = result["count"]
-            deleted_ids = result["ids"]
-
-        # Clean up tracked file entries pointing to deleted saves
-        files_state = save_state.get("files", {})
-        if deleted_ids:
-            to_remove = [fn for fn, fs in files_state.items() if fs.get("tracked_save_id") in deleted_ids]
-            for fn in to_remove:
-                del files_state[fn]
-                cleaned_files += 1
-
-        del slots_dict[slot]
-        self.save_state()
-
-        return {
-            "success": True,
-            "deleted_server_saves": deleted_server_saves,
-            "cleaned_files": cleaned_files,
-        }
+__all__ = ["SaveService"]
